pyproject.toml
[tool.poetry]
name = "midi-generation"
version = "0.1.0"
package-mode = false
description = ""
authors = ["nibura2002 <nibura2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.42.0"
langchain = "^0.3.18"
openai = "^1.61.1"
mido = "^1.3.3"
python-dotenv = "^1.0.1"
autopep8 = "^2.3.2"
langchain-openai = "^0.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
